remove passing of eval state and gc handler
